backend/ml_match_service/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
pymongo
python-dotenv
numpy
scikit-learn
requests
